# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sharefy",
        "description": "",
        "contact": {
            "email": "equip@sharefy.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/follow/user/{username}": {
            "post": {
                "tags": [
                    "follow"
                ],
                "summary": "Seguir a un usuari",
                "description": "Dado un username del usuari a seguir, es crea la relació en la taula followers.",
                "operationId": "App\\Http\\Controllers\\Api\\FollowerDataController::create",
                "parameters": [
                    {
                        "name": "user_follower",
                        "in": "query",
                        "description": "Username del usuari que fa l'acció de seguir a un altre usuari.",
                        "required": true
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'Has començat a seguir a {username} =P'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "401": {
                        "description": "Devuelve un json con el error: 'Restricció: un usuari no pot seguir-se a ell mateix'."
                    },
                    "404": {
                        "description": "Devuelve un json con el error: 'Usuari no trobat'."
                    }
                }
            }
        },
        "/api/follow/user/{follower}/user/{followed}": {
            "get": {
                "tags": [
                    "follow"
                ],
                "summary": "Comprova si un usuari segueix a un altre.",
                "description": "Dado un username del seguidor {follower} i un altre username {followed}, comprova si el segueix o no.",
                "operationId": "App\\Http\\Controllers\\Api\\FollowerDataController::is_following",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el value: 'true' o 'false'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "401": {
                        "description": "Devuelve un json con el error: 'Restricció: un usuari no pot seguir-se a ell mateix."
                    },
                    "404": {
                        "description": "Devuelve un json con el error: 'Usuari no trobat'."
                    }
                }
            },
            "delete": {
                "tags": [
                    "follow"
                ],
                "summary": "Deixar de seguir a un usuari",
                "description": "Dado un username del seguidor {follower} i un username d'usuari que seguix {followed}, s'esborra la relació de followers.",
                "operationId": "App\\Http\\Controllers\\Api\\FollowerDataController::delete",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json con el mensaje: 'Has deixat de seguir a {username} =G'."
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "401": {
                        "description": "Devuelve un json con el error: 'Restricció: un usuari no pot deixar de seguir-se a ell mateix'."
                    },
                    "402": {
                        "description": "Devuelve un json con el error: 'No pots deixar de seguir a cap usuari que encara no segueix'."
                    },
                    "404": {
                        "description": "Devuelve un json con el error: 'Usuari no trobat'."
                    }
                }
            }
        },
        "/api/user/{id}/followers": {
            "get": {
                "tags": [
                    "follow"
                ],
                "summary": "Dado un id d'un usuari et retorna els seus seguidors.",
                "description": "Dado un id usuari et retorna un json amb les dades dels seus seguidors {id, username, photo_path}.",
                "operationId": "App\\Http\\Controllers\\Api\\FollowerDataController::get_followers",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json: followers: [ {id, username, photo_path} ]"
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "404": {
                        "description": "Devuelve un json con el missage: 'Aquest usuari no existeix o no té cap seguidor'."
                    }
                }
            }
        },
        "/api/user/{id}/followed": {
            "get": {
                "tags": [
                    "follow"
                ],
                "summary": "Dado un id d'un usuari et retorna els usuaris que segueix.",
                "description": "Dado un id usuari et retorna un json amb les dades dels usuaris que segueix {id, username, photo_path}.",
                "operationId": "App\\Http\\Controllers\\Api\\FollowerDataController::get_followed",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Valor del token_access.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un json: followed: [ {id, username, photo_path} ]"
                    },
                    "400": {
                        "description": "Devuelve un json con el error: 'error en els paràmetres'."
                    },
                    "404": {
                        "description": "Devuelve un json con el missage: 'Aquest usuari no existeix o no segueix a cap usuari'."
                    }
                }
            }
        },
...
    },
...
}